fix: adjust padding, add shadows to dropdowns
