+++ src/assets/styles/abstracts/_variables.scss
@@ -71,9 +71,4 @@ $scale: (
     }
 
     --max-content: #{(rem(1234))};
-
-    // Navigation
-    --nav-current-background: transparent;
-    --nav-submenu-background: var(--nav-focus-background);
-    --nav-submenu-color: var(--nav-focus-color);
 }
