+++ src/assets/styles/layout/_footer.scss
@@ -7,7 +7,8 @@
     width: 100%;
 
     .wrapper {
-        max-width: var(--max-content);
+        padding-left: var(--scale-4);
+        padding-right: var(--scale-4);
     }
 }
