+++ src/assets/styles/layout/_header.scss
@@ -2,6 +2,11 @@
 // This file contains all styles related to the header of a specific page.
 // -----------------------------------------------------------------------------
 
+.header .wrapper {
+    padding-left: var(--scale-4);
+    padding-right: var(--scale-4);
+}
+
 .header h1,
 .header .subtitle {
     @media (min-width: 1024px) {
